+++ src/components/features/widgets/course/CourseListWithBookmarkMutate/index.tsx
@@ -1,6 +1,7 @@
-import useDeleteBookmarkMutation from "@/hooks/feature/query/useDeleteBookmarkMutation";
 import type { CourseDifficulty } from "@/types/course";
-import useBookmarkMutation from "@hooks/feature/query/useBookmarkMutation";
+import useBookmarkMutation from "@hooks/feature/query/mutate/useBookmarkMutation";
+import useDeleteBookmarkMutation from "@hooks/feature/query/mutate/useDeleteBookmarkMutation";
+import useCoursePathwayPrefetch from "@hooks/feature/query/prefetch/useCoursePathwayPrefetch";
 import CourseList from "@shared/ui/CourseList";
 
 interface CourseListWithBookmarkMutateProps {
@@ -21,6 +22,7 @@ export default function CourseListWithBookmarkMutate({
   image,
   ...props
 }: CourseListWithBookmarkMutateProps) {
+  useCoursePathwayPrefetch({ courseId: id });
   const { mutate: postBookmark } = useBookmarkMutation();
   const { mutate: deleteBookmark } = useDeleteBookmarkMutation();
