+++ src/hooks/feature/query/query/useBasesDetailQuery/index.ts
