+++ src/hooks/feature/query/query/useCoursePathwayQuery/index.ts
